+++ src/lib/numeric/interpolation/Interpolator.cpp
@@ -6,7 +6,7 @@ using namespace gcm;
 template<class TVector>
 void Interpolator<TVector>::minMaxInterpolate(TVector &res, std::vector<TVector> &src, const real &q) const {
 	// where is the point to interpolate
-	int k = (int) q;
+	unsigned long k = (unsigned long) q;
 	// check that we perform interpolation, not extrapolation
 	assert_le(k, src.size() - 1);
 	assert_ge(q, 0);
@@ -32,11 +32,11 @@ template<class TVector>
 void Interpolator<TVector>::interpolate(TVector &res, std::vector<TVector> &src, const real &q) const {
 	// Newton interpolation
 	res = src[0];
-	const int p = src.size() - 1; // order of interpolation
+	const int p = (int)src.size() - 1; // order of interpolation
 	for (int i = 1; i <= p; i++) {
 		for (int j = 0; j < p - i + 1; j++) {
-			// TODO - make all this linal operations faster
-			src[j] = (src[j + 1] - src[j]) /* recurrent finite differences */
+			// TODO - make all this linal operations faster, replace std::vector
+			src[(unsigned long)j] = (src[(unsigned long)j + 1] - src[(unsigned long)j]) /* recurrent finite differences */
 			         * ((q - i + 1) / i) /* coordinate and factorial */ ;
 		}
 		res += src[0];
